requirements.txt
# Core
streamlit>=1.32.0
langchain>=0.1.17
langchain-community>=0.0.30
langchain-core>=0.1.52
langchain-google-genai>=0.0.9
langchain-text-splitters>=0.0.1
google-generativeai>=0.3.2

# Embeddings
sentence-transformers>=2.2.2
huggingface-hub>=0.23.0

# File loading
unstructured>=0.12.5
docx2txt>=0.8
PyMuPDF>=1.23.3  # Required by PyPDFLoader
pdfminer.six>=20221105  # Alternative PDF backend if needed

# Utils
python-dotenv>=1.0.1
tqdm>=4.66.4

# Optional: For typing and hashing
pydantic>=2.7.1
